+++ packages/domain-event/src/models/requirement.model.ts
@@ -24,6 +24,9 @@ export class RequirementModel {
   @Column({ type: 'uuid', nullable: true })
   createdBy?: string;
 
+  @Column({ type: 'uuid', nullable: true })
+  updatedBy?: string = undefined;
+
   @ManyToMany(() => EventModel, (event) => event.requirements)
   events?: EventModel[];
 }