+++ packages/domain-event/src/test/__test-utils__/mocks/event.repository.mock.ts
@@ -6,6 +6,7 @@ export const createEventRepositoryMock = (): jest.Mocked<IEventRepository> =>
     findAll: jest.fn(),
     findById: jest.fn(),
     create: jest.fn(),
+    createWithEventCreated: jest.fn(),
     update: jest.fn(),
     delete: jest.fn(),
     search: jest.fn(),