+++ packages/outbox/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.8.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/database@3.0.0
+  - @volontariapp/messaging@2.1.1
+
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.1.0
+
 ## 0.8.0
 
 ### Minor Changes